fix(sonar): dedupe owner meter subject helpers for PR quality gate
